Editor: Fix Addon Editor not loading when CustomCSS disabled
